Change CNS to OCS (#331)
I changed names from "Container-native Storage" to "OpenShift Container Storage" and changed "CNS" to "OCS". There are several places that I did not change because I wasn't sure if doing so would mess up the flow of the lab. Also, when it specifically called out version 3.9, I left it as CNS because I believe the OCS name is specific to version 3.10