Update infra-nodes.adoc
"“The use of taints/tolerations is not covered in any of these exercises.” → untrue, there is a section on them!" - Removed
"First use of ‘taints and tolerations’ here - many will not know what these terms mean. Suggest saying (in 1 sentence) what it they are and that they’ll be explored later on)" - Added
"“Logging - OpenShift’s log aggregation solution is not installed by default. There is a dedicated lab exercise that goes through the configuration and deployment of logging.”
This feels unnecessary to have here. The next section is all about logging."
-removed