Latest batch of appmod updates (#641)
* Update installation.adoc

- “new installer-provisioned infrastructure” - not new anymore."
Removed "new" as requested

- “you need to type it in” - I was able to manually copy/paste, suggest wording here to that effect - typing it in sucks." 
Changed from "type it" to "Copy/paste it"

"You’ll notice that there is a 4-digit alphanumeric string (eg: f4a3) in the hostname. This string is your GUID. Since you will often use GUID, it makes sense to export it as an environment variable” The GUID is already set in the environment (and is 5 digits, not 4)."
 - Changed from 4 to 5, replaced example to 5 as well. Wasn't sure if we should get rid of the step completely since he said it is already set by the environment so I have just left it as it is.

* Update app-mgmt-basics.adoc

- "There is mention of Dockerhub here, but it’s Quay - recommend not mentioning Dockerhub anymore"
Replaced Mentioning of "Dockerhub" with "Quay" 
-belog -> belong (misspelling)
Replaced

* Update infra-nodes.adoc

"“The use of taints/tolerations is not covered in any of these exercises.” → untrue, there is a section on them!" - Removed
"First use of ‘taints and tolerations’ here - many will not know what these terms mean. Suggest saying (in 1 sentence) what it they are and that they’ll be explored later on)" - Added
"“Logging - OpenShift’s log aggregation solution is not installed by default. There is a dedicated lab exercise that goes through the configuration and deployment of logging.”
This feels unnecessary to have here. The next section is all about logging."
-removed

* Update ldap-groupsync.adoc

“The installer configured a Route for Prometheus by default. Go ahead and control+click the Prometheus link to open it in your browser. You’ll be greeted with a login screen. Click the Log in with OpenShift button, then select the ldap auth mechanism, and use the fancyuser1 user that you gave cluster-reader privileges to earlier.” 
→ It doesn’t tell me what the password is for logging into openshift as fancyuser1 - I had to scroll back and find a place where it shows the “oc login” command to figure out the password. You should call out the password here (perhaps making it copy/pasteable)
-- Moved to text box format for password to separate it from regular block of text as requested

* Update clusterresourcequota.adoc

accross -> across
- Spelling change

* Update app-mgmt-basics.adoc

Change the oc get rs to look more like what the terminal actually produces.

* Update machinesets.adoc

Since we only have 2 workers and scale up to three later I am changing this to match

* Update installation.adoc

Lowered to 2 workers in text

* Update environment.adoc

lower to 2 worker nodes

* Update monitoring-basics.adoc

- Added extra mention to not use the built-in console and to use the external console.

-Updated description of notification's.

* Update ldap-groupsync.adoc

Updating to clickable link - also stating not to use the built-in console as requested

* Update clusterresourcequota.adoc

updated so it didn't show example with project not yet created as requested

* Update installation.adoc

updated workers list to include only 2 workers

* Create settings_app_mod.sh

Setting for app mod

* Update logging.adoc

- Added warnings about using the built-in console
-Added further warnings to make sure they change the YAML file before proceeding

* Update ldap-groupsync.adoc

- add login for oc login with kubeadmin
- warning to wait if you receive error

* Update clusterresourcequota.adoc

message about deprecation warnings.

* adding ACS-content

Co-authored-by: jnewsome97 <50967636+jnewsome97@users.noreply.github.com>
Co-authored-by: mfoster_stackrox <michael.foster@stackrox.com>